// packages/web/src/shared/api/generated/model/debtAccountDto.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * Finnn API
 * Backend API for Finnn finance tracking
 * OpenAPI spec version: 0.1.0
 */

export interface DebtAccountDto {
  id: string;
  name: string;
  currency: string;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  icon?: string | null;
}